Working on Proccessor, log and k8s processor
