+++ modules/framework-k8s/src/main/java/it/smartcommunitylabdhub/framework/k8s/processors/K8sProcessor.java
@@ -1,0 +1,30 @@
+package it.smartcommunitylabdhub.framework.k8s.processors;
+
+import it.smartcommunitylabdhub.commons.annotations.common.RunProcessorType;
+import it.smartcommunitylabdhub.commons.infrastructure.RunProcessor;
+import it.smartcommunitylabdhub.commons.infrastructure.RunRunnable;
+import it.smartcommunitylabdhub.commons.models.entities.run.Run;
+import it.smartcommunitylabdhub.commons.models.entities.run.RunBaseStatus;
+import it.smartcommunitylabdhub.framework.k8s.runnables.K8sRunnable;
+import org.springframework.stereotype.Component;
+
+import java.io.Serializable;
+import java.util.Map;
+
+@RunProcessorType(stages = {"onRunning", "onComplete", "onError", "onStopped"}, id = "k8sProcessor")
+@Component("k8sProcessor")
+public class K8sProcessor implements RunProcessor<RunBaseStatus> {
+    @Override
+    public RunBaseStatus process(Run run, RunRunnable runRunnable, RunBaseStatus status) {
+        if (runRunnable instanceof K8sRunnable) {
+            Map<String, Serializable> res = ((K8sRunnable) runRunnable).getResults();
+            //extract k8s details
+
+//            return PythonRunStatus.builder().k8s(res).build();
+
+            return null;
+        }
+
+        return null;
+    }
+}
