+++ modules/runtime-python/src/main/java/it/smartcommunitylabdhub/runtime/python/PythonRuntime.java
@@ -62,8 +62,6 @@ public class PythonRuntime implements Runtime<PythonFunctionSpec, PythonRunSpec,
 
     public static final String RUNTIME = "python";
 
-    //TODO make configurable
-    public static final int MAX_METRICS = 300;
 
     @Autowired
     private SecretService secretService;
@@ -293,22 +291,22 @@ public class PythonRuntime implements Runtime<PythonFunctionSpec, PythonRunSpec,
     @Override
     public PythonRunStatus onRunning(Run run, RunRunnable runnable) {
         //extract info for status
-        if (runnable instanceof K8sRunnable) {
-            Map<String, Serializable> res = ((K8sRunnable) runnable).getResults();
-            //extract k8s details
-            //TODO
-
-            //extract logs
-            List<CoreLog> logs = ((K8sRunnable) runnable).getLogs();
-            List<CoreMetric> metrics = ((K8sRunnable) runnable).getMetrics();
-
-            if (logs != null) {
-                writeLogs(run, logs, metrics);
-            }
-
-            //dump as-is
-            return PythonRunStatus.builder().k8s(res).build();
-        }
+//        if (runnable instanceof K8sRunnable) {
+//            Map<String, Serializable> res = ((K8sRunnable) runnable).getResults(); // TODO k8sProcessor.
+//            //extract k8s details
+//
+//
+////            //extract logs
+////            List<CoreLog> logs = ((K8sRunnable) runnable).getLogs();
+////            List<CoreMetric> metrics = ((K8sRunnable) runnable).getMetrics();
+////
+////            if (logs != null) {
+////                writeLogs(run, logs, metrics);
+////            }
+//
+//            //dump as-is
+//            return PythonRunStatus.builder().k8s(res).build();
+//        }
 
         return null;
     }
@@ -346,7 +344,7 @@ public class PythonRuntime implements Runtime<PythonFunctionSpec, PythonRunSpec,
             List<CoreMetric> metrics = ((K8sRunnable) runnable).getMetrics();
 
             if (logs != null) {
-                writeLogs(run, logs, metrics);
+//                writeLogs(run, logs, metrics);
             }
 
             //dump as-is
@@ -369,7 +367,7 @@ public class PythonRuntime implements Runtime<PythonFunctionSpec, PythonRunSpec,
             List<CoreMetric> metrics = ((K8sRunnable) runnable).getMetrics();
 
             if (logs != null) {
-                writeLogs(run, logs, metrics);
+//                writeLogs(run, logs, metrics);
             }
 
             //dump as-is
@@ -392,7 +390,7 @@ public class PythonRuntime implements Runtime<PythonFunctionSpec, PythonRunSpec,
             List<CoreMetric> metrics = ((K8sRunnable) runnable).getMetrics();
 
             if (logs != null) {
-                writeLogs(run, logs, metrics);
+//                writeLogs(run, logs, metrics);
             }
 
             //dump as-is
@@ -431,156 +429,4 @@ public class PythonRuntime implements Runtime<PythonFunctionSpec, PythonRunSpec,
         return null;
     }
 
-    private void writeLogs(Run run, List<CoreLog> logs, List<CoreMetric> metrics) {
-        String runId = run.getId();
-        Instant now = Instant.now();
-
-        //logs are grouped by pod+container, search by run and create/append
-        Map<String, Log> entries = logService
-            .getLogsByRunId(runId)
-            .stream()
-            .map(e -> {
-                K8sLogStatus status = new K8sLogStatus();
-                status.configure(e.getStatus());
-
-                String pod = status.getPod() != null ? status.getPod() : "";
-                String container = status.getContainer() != null ? status.getContainer() : "";
-                String namespace = status.getNamespace() != null ? status.getNamespace() : "";
-                String key = namespace + pod + container;
-
-                if (StringUtils.hasText(runId)) {
-                    return Map.entry(key, e);
-                } else {
-                    return null;
-                }
-            })
-            .filter(e -> e != null)
-            .collect(Collectors.toMap(e -> e.getKey(), e -> e.getValue()));
-
-        //reformat metrics grouped per container
-        //TODO refactor
-        Map<String, HashMap<String, Serializable>> mmetrics = new HashMap<>();
-        if (metrics != null) {
-            metrics.forEach(m -> {
-                if (m.metrics() != null) {
-                    m
-                        .metrics()
-                        .forEach(cm -> {
-                            String key = m.namespace() + m.pod() + cm.getName();
-                            if (cm.getUsage() != null) {
-                                HashMap<String, String> usage = cm
-                                    .getUsage()
-                                    .entrySet()
-                                    .stream()
-                                    .collect(
-                                        Collectors.toMap(
-                                            e -> e.getKey(),
-                                            e -> e.getValue().toSuffixedString(),
-                                            (prev, next) -> next,
-                                            HashMap::new
-                                        )
-                                    );
-
-                                HashMap<String, Serializable> mm = new HashMap<>();
-                                mm.put("timestamp", m.timestamp());
-                                mm.put("window", m.window());
-                                mm.put("usage", usage);
-                                mmetrics.put(key, mm);
-                            }
-                        });
-                }
-            });
-        }
-
-        logs.forEach(l -> {
-            try {
-                String key = l.namespace() + l.pod() + l.container();
-
-                if (entries.get(key) != null) {
-                    //update
-                    Log log = entries.get(key);
-                    log.setContent(l.value());
-
-                    //check if metric is available
-                    if (mmetrics.containsKey(key)) {
-                        HashMap<String, Serializable> metric = mmetrics.get(key);
-
-                        //append to status
-                        K8sLogStatus logStatus = new K8sLogStatus();
-                        logStatus.configure(log.getStatus());
-
-                        List<Serializable> list = logStatus.getMetrics() != null
-                            ? new ArrayList<>(logStatus.getMetrics())
-                            : new ArrayList<>();
-
-                        list.addLast(metric);
-                        logStatus.setMetrics(list);
-
-                        //check if we need to slice
-                        //TODO cleanup
-                        if (list.size() > MAX_METRICS) {
-                            Collections.reverse(list);
-                            List<Serializable> slice = new ArrayList<>(list.subList(0, MAX_METRICS));
-                            Collections.reverse(slice);
-                            logStatus.setMetrics(slice);
-                        }
-
-                        log.setStatus(logStatus.toMap());
-                    }
-
-                    logService.updateLog(log.getId(), log);
-                } else {
-                    //add as new
-                    LogSpec logSpec = new LogSpec();
-                    logSpec.setRun(runId);
-                    logSpec.setTimestamp(now.toEpochMilli());
-
-                    K8sLogStatus logStatus = new K8sLogStatus();
-                    logStatus.setPod(l.pod());
-                    logStatus.setContainer(l.container());
-                    logStatus.setNamespace(l.namespace());
-
-                    //check if metric is available
-                    if (mmetrics.containsKey(key)) {
-                        HashMap<String, Serializable> metric = mmetrics.get(key);
-
-                        //append to status
-                        List<Serializable> list = logStatus.getMetrics() != null
-                            ? new ArrayList<>(logStatus.getMetrics())
-                            : new ArrayList<>();
-                        list.addLast(metric);
-                        logStatus.setMetrics(list);
-
-                        //check if we need to slice
-                        //TODO cleanup
-                        if (list.size() > MAX_METRICS) {
-                            Collections.reverse(list);
-                            List<Serializable> slice = new ArrayList<>(list.subList(0, MAX_METRICS));
-                            Collections.reverse(slice);
-                            logStatus.setMetrics(slice);
-                        }
-                    }
-
-                    Log log = Log
-                        .builder()
-                        .project(run.getProject())
-                        .spec(logSpec.toMap())
-                        .status(logStatus.toMap())
-                        .content(l.value())
-                        .build();
-
-                    logService.createLog(log);
-                }
-            } catch (
-                NoSuchEntityException
-                | IllegalArgumentException
-                | SystemException
-                | BindException
-                | DuplicatedEntityException e
-            ) {
-                //invalid, skip
-                //TODO handle
-            }
-        });
-    }
 }
